chore: moved ui to playlist entity
