--- components/playlist-form/index.ts
@@ -1,3 +1,0 @@
-export * from "./playlist-form";
-export * from "./playlist-form-modal";
-export * from "./playlist-form-track";
