--- features/playlist/playlist-dto.ts
@@ -1,10 +1,0 @@
-import { Playlist } from "./types";
-
-/**
- * @deprecated
- */
-
-export interface CreatePlayListDto extends Playlist {}
-export interface UpdatePlayListDto extends Playlist {
-  id: number;
-}
