--- features/playlist/types.ts
@@ -1,9 +1,0 @@
-export type Playlist = {
-  id?: number;
-  name: string;
-  isActive: boolean;
-  comments: unknown[];
-  creatorId: number;
-  creator: unknown;
-  songs: unknown[];
-};
